src/requirements.txt
argparse
boto3
youtube_transcript_api
streamlit
pyyaml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
streamlit-browser-language